+++ sdk/src/main/java/com/mapbox/maps/NativeMapProvider.java
@@ -1,7 +1,0 @@
-package com.mapbox.maps;
-
-class NativeMapProvider {
-  static Map getNativeMap(MapView mapView) {
-    return MapProvider.INSTANCE.getNativeMapCore(mapView);
-  }
-}