+++ sdk/src/test/java/com/mapbox/maps/ShadowMapRecorder.java
@@ -1,0 +1,7 @@
+package com.mapbox.maps;
+
+import org.robolectric.annotation.Implements;
+
+@Implements(MapRecorder.class)
+public class ShadowMapRecorder {
+}